requirements.txt
beautifulsoup4
requests
pytest
dogpile.cache
